+++ CHANGELOG.md
@@ -22,3 +22,8 @@ This is the changelog of the mediasoup demo application. To check the changes in
 ## 2019-01-15
 
 * Update mediasoup to 2.6.6. It fixes an important [port leak bug](https://github.com/versatica/mediasoup/issues/259).
+
+
+## 2019-03-21
+
+* Always ask mic to avoid browsers' autoplay policy when in spy mode.